Found reason for failing test and added another
Test was failing due to:
* Inproper expected generation logic in Scenerio
* No concept of date/time in V1 files

Generation logic has been patually fixed. Wrote test for date and time
so it can be added next time. The idea is that date and time can be read
from the modify timestamp of the audio files. Then the timestamp could
be appended to the end of the V1 index line seperated by a pipe (|).
The parser should be backwards compatible with old index files that do
not include the date time information.